refactor(visualization): custom config dialog
