+++ visualization/app/codeCharta/ui/customConfigs/addCustomConfigButton/addCustomConfigDialog/addCustomConfigDialog.component.scss
@@ -1,0 +1,5 @@
+mat-dialog-content {
+    mat-form-field {
+        width: 100%;
+    }
+}
