--- visualization/app/codeCharta/ui/customConfigs/addCustomConfigButton/addCustomConfigDialog/addCustomDialog.component.scss
@@ -1,61 +1,0 @@
-@use "variables";
-
-.cc-add-custom-config-dialog .mat-mdc-dialog-container {
-    .dialog-content {
-        mat-form-field {
-            width: 100%;
-
-            &.custom-config-note {
-                padding-top: 30px;
-            }
-        }
-
-        label {
-            color: rgba(0, 0, 0, 0.54);
-            font-size: 13px;
-        }
-
-        input {
-            margin-top: 12px;
-        }
-    }
-
-    .mat-mdc-dialog-actions {
-        margin: 0;
-        padding: 0 8px;
-        position: relative;
-
-        cc-download-and-purge-configs {
-            text-align: right;
-
-            .exceed-local-storage-message {
-                display: block;
-                text-align: left;
-                padding: 0 16px;
-                color: #b02929;
-            }
-            .exceed-local-storage-button {
-                margin-right: 89px;
-            }
-        }
-
-        .mat-mdc-button {
-            padding: 0 8px;
-            margin: 8px 0 8px 8px;
-
-            &.exceed-local-storage-button {
-                color: #ffffff;
-                background-color: #b02929;
-            }
-
-            &.add-custom-config-button {
-                color: variables.$cc-primary-color;
-                background-color: rgba(138, 132, 132, 0.08);
-            }
-
-            &.add-custom-config-button[disabled] {
-                color: grey;
-            }
-        }
-    }
-}
