+++ visualization/app/codeCharta/ui/customConfigs/addCustomConfigButton/addCustomConfigDialog/downloadAndPurgeConfigs/downloadAndPurgeConfigs.component.scss
@@ -1,0 +1,16 @@
+:host {
+    text-align: right;
+}
+
+.exceed-local-storage-message {
+    display: block;
+    text-align: left;
+    padding: 0 16px;
+    color: #b02929;
+}
+
+.exceed-local-storage-button {
+    margin-right: 89px;
+    color: #ffffff;
+    background-color: #b02929;
+}
